Support TLSv1.2 by default under Java 7
Following changes in Maven Central, only TLSv1.2 is accepted.
This change turns on TLSv1, TLSv1.1 and TLSv1.2 for all connections from
Gradle when the Java version is 7.
As documented in https://docs.oracle.com/javase/7/docs/technotes/guides/security/SunProviders.html
in the Protocols section, this is safe, except for versions before 7u75
that were relying on SSLv3.
However that last protocol is really broken and no one should be using
it. And we definitely cannot turn it on here.

It also adds that property to tests that do interact with Maven Central.